consolidate trace code
