+++ pydl/pydlutils/tests/t/sdss_traceset.fits
