--- pydl/pydlutils/trace/__init__.py
@@ -1,12 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the trace directory of idlutils.
-"""
-from .fchebyshev import fchebyshev
-from .fchebyshev_split import fchebyshev_split
-from .fpoly import fpoly
-from .func_fit import func_fit
-from .TraceSet import TraceSet
-from .traceset2xy import traceset2xy
-from .xy2traceset import xy2traceset
